# openapi/openapi.yaml
openapi: '3.1.1'
info:
  title: ViaTrade Web API
  description: API для платформы инвестиционного анализа ViaTrade
  version: v1
paths:
  /api/v1/instruments/statistics:
    get:
      tags:
        - Instruments
      operationId: GetInstrumentStatistics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentStatisticsResult'
        '400':
          description: Bad Request - Invalid input data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/400
                title: Bad Request
                status: 400
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '401':
          description: Unauthorized - Authentication is required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/401
                title: Unauthorized
                status: 401
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '403':
          description: Forbidden - Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/403
                title: Forbidden
                status: 403
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '404':
          description: Not Found - Resource does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/404
                title: Not Found
                status: 404
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '408':
          description: Request Timeout - Operation timed out
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/408
                title: Request Timeout
                status: 408
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '409':
          description: Conflict - Resource state conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/409
                title: Conflict
                status: 409
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '422':
          description: Unprocessable Content - Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/422
                title: Unprocessable Content
                status: 422
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '500':
          description: Internal Server Error - Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/500
                title: Internal Server Error
                status: 500
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '503':
          description: Service Unavailable - Temporary infrastructure issue
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/503
                title: Service Unavailable
                status: 503
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
      security:
        - Bearer: [ ]
  /api/v1/instruments:
    get:
      tags:
        - Instruments
      operationId: GetInstruments
      parameters:
        - name: ticker
          in: query
          schema:
            type: string
        - name: searchText
          in: query
          schema:
            maxLength: 100
            type: string
        - name: page
          in: query
          schema:
            maximum: 21474837
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            maximum: 100
            minimum: 1
            type: integer
            format: int32
            default: 20
        - name: sortBy
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/InstrumentSortField'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentResultPageResult'
        '400':
          description: Bad Request - Invalid input data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/400
                title: Bad Request
                status: 400
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '401':
          description: Unauthorized - Authentication is required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/401
                title: Unauthorized
                status: 401
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '403':
          description: Forbidden - Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/403
                title: Forbidden
                status: 403
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '404':
          description: Not Found - Resource does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/404
                title: Not Found
                status: 404
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '408':
          description: Request Timeout - Operation timed out
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/408
                title: Request Timeout
                status: 408
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '409':
          description: Conflict - Resource state conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/409
                title: Conflict
                status: 409
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '422':
          description: Unprocessable Content - Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/422
                title: Unprocessable Content
                status: 422
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '500':
          description: Internal Server Error - Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/500
                title: Internal Server Error
                status: 500
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '503':
          description: Service Unavailable - Temporary infrastructure issue
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/503
                title: Service Unavailable
                status: 503
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
      security:
        - Bearer: [ ]
  '/api/v1/instruments/{instrumentId}':
    get:
      tags:
        - Instruments
      operationId: GetInstrumentById
      parameters:
        - name: instrumentId
          in: path
          required: true
          schema:
            exclusiveMinimum: 0
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentResult'
        '400':
          description: Bad Request - Invalid input data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/400
                title: Bad Request
                status: 400
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '401':
          description: Unauthorized - Authentication is required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/401
                title: Unauthorized
                status: 401
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '403':
          description: Forbidden - Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/403
                title: Forbidden
                status: 403
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '404':
          description: Not Found - Resource does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/404
                title: Not Found
                status: 404
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '408':
          description: Request Timeout - Operation timed out
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/408
                title: Request Timeout
                status: 408
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '409':
          description: Conflict - Resource state conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/409
                title: Conflict
                status: 409
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '422':
          description: Unprocessable Content - Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/422
                title: Unprocessable Content
                status: 422
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '500':
          description: Internal Server Error - Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/500
                title: Internal Server Error
                status: 500
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '503':
          description: Service Unavailable - Temporary infrastructure issue
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/503
                title: Service Unavailable
                status: 503
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
      security:
        - Bearer: [ ]
  '/api/v1/instruments/{instrumentId}/strategies':
    get:
      tags:
        - Instruments
      operationId: GetStrategiesByInstrument
      parameters:
        - name: instrumentId
          in: path
          required: true
          schema:
            exclusiveMinimum: 0
            type: integer
            format: int32
        - name: name
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/StrategySortField'
        - name: page
          in: query
          schema:
            maximum: 21474837
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            maximum: 100
            minimum: 1
            type: integer
            format: int32
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StrategySubscriptionResultPageResult'
        '400':
          description: Bad Request - Invalid input data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/400
                title: Bad Request
                status: 400
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '401':
          description: Unauthorized - Authentication is required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/401
                title: Unauthorized
                status: 401
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '403':
          description: Forbidden - Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/403
                title: Forbidden
                status: 403
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '404':
          description: Not Found - Resource does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/404
                title: Not Found
                status: 404
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '408':
          description: Request Timeout - Operation timed out
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/408
                title: Request Timeout
                status: 408
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '409':
          description: Conflict - Resource state conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/409
                title: Conflict
                status: 409
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '422':
          description: Unprocessable Content - Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/422
                title: Unprocessable Content
                status: 422
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '500':
          description: Internal Server Error - Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/500
                title: Internal Server Error
                status: 500
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '503':
          description: Service Unavailable - Temporary infrastructure issue
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/503
                title: Service Unavailable
                status: 503
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
      security:
        - Bearer: [ ]
...
  /api/v1/signals/statistics:
    get:
      tags:
        - Signals
      operationId: GetStatistics
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalStatisticsResult'
        '400':
          description: Bad Request - Invalid input data
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/400
                title: Bad Request
                status: 400
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '401':
          description: Unauthorized - Authentication is required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/401
                title: Unauthorized
                status: 401
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '403':
          description: Forbidden - Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/403
                title: Forbidden
                status: 403
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '404':
          description: Not Found - Resource does not exist
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/404
                title: Not Found
                status: 404
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '408':
          description: Request Timeout - Operation timed out
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/408
                title: Request Timeout
                status: 408
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '409':
          description: Conflict - Resource state conflict
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/409
                title: Conflict
                status: 409
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '422':
          description: Unprocessable Content - Business rule violation
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/422
                title: Unprocessable Content
                status: 422
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '500':
          description: Internal Server Error - Unexpected error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/500
                title: Internal Server Error
                status: 500
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
        '503':
          description: Service Unavailable - Temporary infrastructure issue
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                type: https://httpstatuses.io/503
                title: Service Unavailable
                status: 503
                detail: Safe error description.
                instance: /api/resource
                code: error_code
                traceId: 0HNABC123:00000001
      security:
        - Bearer: [ ]
...
components:
  schemas:
...
    InstrumentResult:
      required:
        - id
        - ticker
      type: object
      properties:
        id:
          type: integer
          format: int32
        ticker:
          type: string
        description:
          type: string
      additionalProperties: false
    InstrumentResultPageResult:
      required:
        - items
        - page
        - pageSize
        - totalCount
        - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InstrumentResult'
        totalCount:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
          readOnly: true
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
...
    InstrumentSortField:
      enum:
        - tickerAsc
        - tickerDesc
      type: string
    InstrumentStatisticsResult:
      required:
        - totalInstruments
      type: object
      properties:
        totalInstruments:
          type: integer
          format: int32
      additionalProperties: false
...
    ProblemDetails:
      required:
        - type
        - title
        - status
        - detail
        - instance
        - code
        - traceId
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      example:
        type: https://httpstatuses.io/400
        title: Validation Failed
        status: 400
        detail: One or more validation errors occurred.
        instance: /api/trades
        code: validation_failed
        traceId: 0HNABC123:00000001
...
    SignalStatisticsResult:
      required:
        - buySignals
        - sellSignals
        - totalSignals
      type: object
      properties:
        totalSignals:
          type: integer
          format: int32
        buySignals:
          type: integer
          format: int32
        sellSignals:
          type: integer
          format: int32
      additionalProperties: false
...
    StrategySortField:
      enum:
        - nameAsc
        - nameDesc
        - accuracyAsc
        - accuracyDesc
      type: string
...
    StrategySubscriptionResultPageResult:
      required:
        - items
        - page
        - pageSize
        - totalCount
        - totalPages
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StrategySubscriptionResult'
        totalCount:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
          readOnly: true
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: Write JWT token without the Bearer prefix.
      scheme: bearer
      bearerFormat: JWT
...